others/crawling/requirements.txt
google-api-python-client==2.108.0
Pillow==10.1.0
python-dotenv==1.0.0 